// PensionManagementSystemUI/src/models/pensioner.ts
export class Pensioner {
    aadhaarNumber?: number;
    accountNumber?: number;
    allowances?: number;
    bankName?: string;
    bankServiceCharge?: number;
    dateOfBirth?: string;
    name?: string;
    panNumber?: string;
    pensionAmount?: number;
    salary?: number;
    typeOfPension?: string;
}